+++ src/lcovReport.ts
@@ -1,92 +1,127 @@
 import * as vscode from "vscode";
-import * as chokidar from "chokidar";
-import { disableDecorations, applyCoverage } from "./decorations";
+import {
+  disableDecorations,
+  applyCoverage,
+  removeDecorationTypes,
+} from "./decorations";
 import path from "path";
 import fs from "fs";
 import { custom } from "./log";
 import { LcovFile } from "lcov-parse";
-import { makePathsAbsolute, readLcovFile } from "./lcov";
-import { config } from "process";
+import { loadLcovFiles } from "./lcov";
+import { updateStatusBar } from "./statusBar";
 
-let lcovFiles: LcovFile[] = [];
+let watcher: fs.FSWatcher | null = null;
 let lcovFilesCache: LcovFile[] = [];
-
-let loaded: boolean = false;
-export let filePath: string = "";
-
-export async function resolveLCOVFile() {
-  if (filePath === "") {
-    return;
-  }
-  lcovFilesCache = lcovFiles;
-  loaded = false;
-  lcovFiles = await readLcovFile(filePath);
-  makePathsAbsolute(lcovFiles);
-  loaded = true;
-}
+let activeLcovFile: string = "";
+let display: boolean = false;
 
 export function getLcovFiles(): LcovFile[] {
-  if (loaded) {
-    return lcovFiles;
-  }
   return lcovFilesCache;
 }
 
-export async function loadFilePath() {
-  let configPath = getFilePath();
-  if (!configPath) {
-    vscode.window.showErrorMessage("Unable read Search Path.");
-    return false;
-  }
-
-  if (!path.isAbsolute(configPath)) {
-    await vscode.workspace.findFiles(configPath).then((files) => {
-      if (files.length > 0) {
-        filePath = files[0].fsPath;
-      } else {
-        vscode.window.showErrorMessage(
-          "Unable to load lcov report: " + configPath
-        );
-        filePath = "";
-      }
-    });
-
-    return;
-  }
-  filePath = configPath;
-}
-
-let watcher: chokidar.FSWatcher | null = null;
-
 export function disableWatchReport() {
-  watcher?.removeAllListeners();
+  if (watcher) {
+    watcher.removeAllListeners();
+  }
   disableDecorations();
 }
 
-export async function activateWatchReport(): Promise<void> {
-  await resolveLCOVFile();
-  fs.watch(filePath, async (event, filename) => {
-    let stop = false;
-    while (!stop) {
-      stop = true;
-      await resolveLCOVFile().then(undefined, (err) => {
-        custom.error("Unable to resolve lcov file");
-        stop = false;
-      });
-    }
-    await applyCoverage();
+export function executeDisplay() {
+  return new Promise<void>((resolve, reject) => {
+    runCoverage().then(
+      () => {
+        display = true;
+        updateStatusBar(display);
+        activateWatchReport();
+        return resolve();
+      },
+      (reason: string) => {
+        vscode.window.showErrorMessage(reason);
+        return reject();
+      }
+    );
   });
 }
 
-function getFilePath(): string {
-  const config = vscode.workspace.getConfiguration("code-coverage-lcov.path");
-  const searchPath: string | undefined = config.get("searchPath");
-  if (!searchPath) {
-    return "";
+export function onDidChangeVisibleTextEditors(
+  editors: readonly vscode.TextEditor[]
+) {
+  if (display) {
+    applyCoverage();
   }
-  return searchPath;
+}
+
+export function executeHide() {
+  disableWatchReport();
+  removeDecorationTypes();
+  display = false;
+  updateStatusBar(display);
 }
 
 export function disableReportWatcher() {
-  watcher?.close();
+  if (watcher) {
+    watcher.close();
+  }
+}
+
+function loadLcovFromFileAndApplyCoverage(
+  resolve: (value: void | PromiseLike<void>) => void,
+  reject: (reason?: any) => void
+) {
+  if (!fs.existsSync(activeLcovFile)) {
+    return reject("Unable to open file: " + activeLcovFile);
+  }
+  loadLcovFiles(activeLcovFile).then(
+    (lcovFiles) => {
+      lcovFilesCache = lcovFiles;
+      applyCoverage();
+    },
+    (reason) => {
+      return reject(reason);
+    }
+  );
+  return resolve();
+}
+
+function runCoverage() {
+  return new Promise<void>((resolve, reject) => {
+    let searchPath = getSearchPath();
+    if (!searchPath) {
+      return reject("Unable read searchPath from config.");
+    }
+
+    // check if file path must be changed to absolute
+    if (!path.isAbsolute(searchPath)) {
+      vscode.workspace.findFiles(searchPath).then((files) => {
+        if (files.length > 0) {
+          activeLcovFile = files[0].fsPath;
+          return loadLcovFromFileAndApplyCoverage(resolve, reject);
+        } else {
+          return reject("Unable to find file in workspace: " + searchPath);
+        }
+      });
+    } else {
+      activeLcovFile = searchPath;
+      return loadLcovFromFileAndApplyCoverage(resolve, reject);
+    }
+  });
+}
+
+function getSearchPath(): string | undefined {
+  const config = vscode.workspace.getConfiguration("code-coverage-lcov.path");
+  return config.get("searchPath");
+}
+
+function activateWatchReport() {
+  watcher = fs.watch(activeLcovFile, async (event, filename) => {
+    custom.log("Event: ", event, " on: ", filename);
+    let promise = new Promise((reject, resolve) => {
+      loadLcovFromFileAndApplyCoverage(reject, resolve);
+    });
+    promise.catch((reason) => {
+      vscode.window.showErrorMessage(reason);
+      executeHide();
+    });
+  });
 }